exer 4.20 & 4.21

+++ part4/blogList/controllers/blogs.js
@@ -2,6 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -18,33 +20,14 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
-
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   try {
     const { title, author, url, likes } = request.body
-
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if(!decodedToken.id){
-      return response.status(401).json({error: 'Token invalid'})
-    }
-
-    const user = await User.findById(decodedToken.id)
-    if (!user) {
-      return response.status(400).json({ error: 'User not found' })
-    }
+    const user = request.user
 
     if (!title || !url) {
       return response.status(400).json({ error: 'title and url are required' })
     }
-
 
     const blog = new Blog({
       title,
@@ -59,16 +42,32 @@ blogsRouter.post('/', async (request, response, next) => {
     await user.save()
 
     const populatedBlog = await savedBlog.populate('user', { username: 1, name: 1 })
-
     response.status(201).json(populatedBlog)
   } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  try {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user.toString() !== user._id.toString()) {
+      return response.status(403).json({
+        error: `Unauthorized: ${user.username} is not the creator of blog '${blog.title}'`
+      })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } catch (error) {
+    next(error)
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
